+++ needle/modules/comms/proxy/proxy_regular.py
@@ -28,7 +28,7 @@ class Module(BaseModule):
     def __init__(self, params):
         BaseModule.__init__(self, params)
         # Setting default output file
-        self.options['output'] = self.local_op.build_output_path_for_file(self, "proxy_regular.out")
+        self.options['output'] = self.local_op.build_output_path_for_file("proxy_regular.out", self)
 
     def module_pre(self):
         return BaseModule.module_pre(self, bypass_app=True)