+++ needle/modules/dynamic/monitor/files.py
@@ -19,7 +19,7 @@ class Module(BackgroundModule):
     def __init__(self, params):
         BackgroundModule.__init__(self, params)
         # Setting defaults
-        self.options['output'] = self.local_op.build_output_path_for_file(self, "modified_files.txt")
+        self.options['output'] = self.local_op.build_output_path_for_file("modified_files.txt", self)
 
     # ==================================================================================================================
     # RUN