+++ CHANGELOG.md
@@ -14,7 +14,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 #### Fixed
 - **[CORE]** Sanitization of parsed plist files
 - **[CORE]** App metadata: show all URI handlers
-- **[CORE]** Invalid characters when parsing plist files _[from @tghosth]_ 
+- **[CORE]** Invalid characters when parsing plist files
 - **[MODULE]** Dump entire keychain _[idea from @tghosth]_
 - **[MODULE]** `storage/caching/screenshot`: OS X support for rendering preview images
 - **[MODULE]** Error saving files in `storage/data/files_*` modules _[from @tghosth]_